# src/web/api/src/main/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WindowsToLinux internal Web API",
    "version": "1.0.0",
    "description": "Phase 5 only. Fixed server-owned internal user/workspace. Loopback, same-origin and explicit headers are not authentication. Never expose publicly. Resource views contain metadata and encrypted references only. Lists are bounded to 1000. Task kinds have closed input schemas; deploy overrides use shared typed field IDs. No shell or client-provided task state is accepted."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8765/api/v1"
    }
  ],
  "paths": {
...
    "/backups": {
      "get": {
        "summary": "List scoped backups",
        "operationId": "get_backups",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Backup"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Safe error; no stack, command, SQL, local path or secret",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/backups/upload": {
      "put": {
        "summary": "Upload and validate portable archive",
        "operationId": "put_backups_upload",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Backup"
                }
              }
            }
          },
          "default": {
            "description": "Safe error; no stack, command, SQL, local path or secret",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Exact http://127.0.0.1:<actual-port>"
          },
          {
            "name": "X-W2L-Client",
            "in": "header",
            "required": true,
            "schema": {
              "const": "web"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/backups/{id}/download": {
      "get": {
        "summary": "Revalidate and download exact archive",
        "operationId": "get_backups_id_download",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "Safe error; no stack, command, SQL, local path or secret",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlationId"
        ]
      },
...
      "Backup": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string"
          },
          "digest": {
            "type": "string"
          },
          "byteCount": {
            "type": "integer",
            "minimum": 0
          },
          "inspection": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "application": {
                "type": "string"
              },
              "schema": {
                "type": "string"
              },
              "canRestore": {
                "type": "boolean"
              },
              "components": {
                "type": "integer",
                "minimum": 0
              },
              "database": {
                "type": "string"
              },
              "provenance": {
                "type": "string"
              }
            },
            "required": [
              "application",
              "schema",
              "canRestore",
              "components",
              "database",
              "provenance"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "createdAt",
          "digest",
          "byteCount",
          "inspection"
        ]
      },
...
    }
  }
}